oimb1y-large-file-processor/repository_after/requirements.txt
Django>=4.2
djangorestframework>=3.14
celery>=5.3
redis>=5.0
django-celery-results>=2.5
python-magic>=0.4.27; sys_platform != 'win32'
python-magic-bin>=0.4.14; sys_platform == 'win32'
boto3>=1.28
django-storages>=1.13
psycopg2-binary>=2.9
Pillow>=10.0
PyPDF2>=3.0
gunicorn>=21.2
pytest>=7.4
pytest-django>=4.5
pytest-cov>=4.1
pytest-env>=1.0
pytest-mock>=3.11
pytest-asyncio>=0.21
coverage>=7.3
Markdown>=3.4
requests>=2.31
mutagen>=1.47
pytest-json-report>=1.5.0